[emacs] Use standard name for my Ruby layer
By using `mjh-ruby` as a layer name, Spacemacs was getting confused and
saying that it had to install the `ruby` layer in order to work with
Ruby files. This clearly isn't the intended behavior, so I went ahead
and renamed my customized layer to `ruby` and deleted the upstream layer
from my installation.

I think I might need to resort to having an entirely hand-crafted set of
layers, but I'm not ready to take that plunge yet.