[git] Add a merged-remote alias
This is useful for finding branches that have been deleted on the
upstream side of their tracking. We use a rebase-over-merge model at
work and it makes it difficult to clean up branches since they do not
show as merged until you rebase them.

This helps fight that problem by showing you all branches that were
deleted on their upstream. That is a useful proxy for branches that you
can remove locally.