[zsh] Allow Zsh to work with Emacs TRAMP
This solution is courtesy of the [Emacs Stack Exchange][1]. Modifying
anything at the `.zshrc` level didn't seem to work, but modifying it at
the `.zshenv` worked a charm.

There are potentially some other modules I want to disable -- notably
`prompt_cr` and `prompt_subst` -- per [this article][2], but it seems to
work as-is.

[1]: https://emacs.stackexchange.com/a/45810
[2]: https://geektivities.wordpress.com/2009/12/03/emacs-with-tramp-using-zsh/