[git] Maintain original commit date when rebasing
An annoyance that I have with rebasing is that it changes the commit
date to the current time (since that's the time the commit happened),
meaning you lose the date on which you were working on a task.

By using the author's date as the commit date, we can work around this
problem and preserve the time history of changes while still using
a rebase workflow.