[emacs] Add a script for easily updating roam
I've found that I don't need to be bothered writing a useful commit
message in my Zettelkasten/org-roam repository and instead am interested
in making it work as quickly as possible. As such, this script first
checks whether we're up-to-date and warns when we are not, then does
a blanket add of files in the repository, commits with the current date,
and pushes to the main remote.

I think that will be enough for now, we'll see if I need to modify it.