[git] Improve diff output for many filetypes
Cribbing this from [Tekin][1], [a follow-up post][2], and [the Git
source][3]. This change allows for [better, more localized output][4] of
diffs when showing a set of changes. In addition, it allows for [seeing
the history of a function/method][5] using the following pattern:

    git log -L :function_name:file_name.ext

That incantation will search for a function or method called
"function_name" in all lines (specified by no numbers prior to the first
colon) in the "file_name.ext" file.

🤯

[1]: https://gist.github.com/tekin/12500956bd56784728e490d8cef9cb81
[2]: https://gist.github.com/ruediger/5647207
[3]: https://github.com/git/git/blob/7f7ebe054af6d831b999d6c2241b9227c4e4e08d/userdiff.c
[4]: https://tekin.co.uk/2020/10/better-git-diff-output-for-ruby-python-elixir-and-more
[5]: https://twitter.com/tgamblin/status/1340834896432926720?s=21