[emacs] Ensure Org settings aren't clobbered
Doom does a very lazy setup of the Org settings in its `:lang org`
module. This is great because it makes it fast to start Emacs, but it
makes it a little more difficult to set settings that conflict with the
module.

This change uses a hook rather than an `after!` macro because that seems
like the lighter weight option.