[emacs] Respect visual line mode with Evil
I find it jarring to have visual lines not respected when moving up and
down through the file. This enables that functionality. I am a little
worried though, since it appears [there are bugs][1] in the
functionality. Though, I believe I used this with Spacemacs for years,
so it should be fine.

[1]: https://github.com/emacs-evil/evil/issues/1168