[emacs] Fix smartparens settings within Org mode
This is a huge hammer of a change, but I can't figure out a better way
to do it.

Tilde completion makes it much easier to write blog posts.

Markdown codeblock completion makes no sense in Org mode.

Single quote pairing is very annoying when I'm in a elisp source block.
These are my three biggest issues for me right now, so I'm keeping the
hacky way of handling it.