[emacs] Add Hugo support to Org mode
I'm in the process of moving my personal site from WordPress to static
files generated by Hugo.

Hugo is nice because it's fast, but it's a little clunky. But the
process of exporting Org to Markdown is very nice with `ox-hugo` so I
want to use this regardless for now.