[emacs] Remove the parenthesis pair for tilde
I thought this would be useful for writing code-related prose since I
use code tags pretty often. And it is. But I also use tilde for other
reasons and it's getting annoying. The benefit is mostly realized on my
work keyboard that has the missing key cap for the grave/tilde key. On
my laptop, I have no trouble hitting it.