[emacs] Use Keychain to manage SSH/GPG
This solves the problem of my ssh-agent not being picked up when I am
attempting to perform Git operations in Magit. I *think* this started
around the time I switched to Emacs 28, but that might be wrong. Anyway,
I use Keychain so I should use it in Emacs as well.

I haven't tested this on the Mac yet, so I don't know if I need to wrap
the config blocks in an `(unless IS-MAC)` so we'll see the next time I
have to restart Emacs.