[git] Remove alias for Flywheel identity
Instead of managing my identity and key on a per-repository basis, I am
now going to use a single key. On my work machine, I have this file set
to use my work identity. My personal machines will default to my open
source identity.

I will still need to manage identity if I work on, for example, Sifter
on my personal machine. But that's easy to do now that I only have to
set the identity and not the key.