[zsh] Switch to Starship as a prompt provider
I _really_ like the aesthetics of the [Pure prompt][1] and have used it
for years. However, since switching to a Yubikey and GPG certificates
for SSH, I keep running into [a bug][2] in the interaction between
pinentry-ncurses and Pure's async functionality. It mangles all future
output for the terminal, launches a spinning pinentry-curses process,
and generally wreaks havoc.

As such, I'm attempting a switch to [Starship][3] as my prompt provider
to see if that alleviates the issue. To replicate the Pure prompt's
functionality, I [patched Starship][4] locally. I intend to push this
upstream after I use it as my daily driver on both Linux and macOS for a
while.

[1]: https://github.com/sindresorhus/pure/
[2]: https://github.com/sindresorhus/pure/issues/366
[3]: https://github.com/starship/starship
[4]: https://github.com/michaelherold/starship/commit/cfb20c57d2325dcc4aa957cad12ba08ac0096eb2