[hidpi] Rely on (X)Wayland scaling
I am currently running [sway-hidpi-git][1] on my laptop that applies [a
patch][2] for managing XWayland scaling at the compositor level. This
works well enough for me except for the cursor size issue outlined in
the pull request. It has proven to be much more stable than the various
environment variable hacks.

It's important to note that the Emacs text scale also requires the GTK
build of Emacs so that it no longer runs under XWayland. I am using the
[emacs-pgtk-native-comp-git][3] package for this.

[1]: https://aur.archlinux.org/packages/sway-hidpi-git/
[2]: https://github.com/swaywm/sway/pull/5090
[3]: https://aur.archlinux.org/packages/emacs-pgtk-native-comp-git/