[ramuh] Add sway configuration
I've been tinkering with this forever and it's to the point where I
would be sad if I lost it. Now that I know about host configs, it seems
appropriate to bring in.

I will consider moving this -- or part of it -- to a tag if I start
using sway elsewhere. Zeromus has an Nvidia card so I can't use it there.