[zsh] Colorize `less` output
Less uses the termcap standard for styling its output, but these values
are not documented anywhere. This partial list [comes from a
StackOverflow answer][1]. I did some source diving a few months ago when
I was originally working through this change but I don't remember what I
learned because I didn't finish the work.

I'm going to give this a try and decide later whether I like it or not.

[1]: https://unix.stackexchange.com/a/108840