[vim] Move to an XDG base directory layout
Vim is one of the few programs that I use that refuses to allow for an
XDG base directory layout. As such, I had to do a bit of hacking to get
it working how I need it to.

I haven't tested this on macOS yet, but the Systemd setup works a treat.