[git] Add master -> main awareness
The trend in Git is to move toward a `main` branch instead of the
`master` branch. To make it easier to work in a world with both names
for the main branch, this change introduces a command that detects the
name of the branch and uses it in relevant commands.