[vim] Update fzf handling to use both repos
I used to only use this on systems where the fzf plugin ended up in the
same place. With the changes to Homebrew on the M1, this ended up
needing more jank, so I opted to just install both repositories. It
won't double-install fzf because it will already be on the path.