[zsh] Reset IFS properly in zprofile
In my effort to repurpose my environment.d files for environment
variables, I change the Internal Field Separator (IFS) to only split by
newlines. This works for my purposes, but I forgot to set it back to the
default (which is, I believe, ` \t\n`). This causes the ZSH completion
dump to become corrupted on Spin.

I figured this out from [this issue on OhMyZsh][1].

[1]: https://github.com/ohmyzsh/ohmyzsh/issues/745#issuecomment-489173934