[git] Ignore extra include from dev
Our internal build tool at Shopify sets a shared include to our Git
config. This leads to an edit to the file whenever I run a command with
the build tool.

This change makes it so Git ignores that line by filtering it from the
commit. On systems where I don't have this build tool, the filter will
be a no-op because it won't exist.