[git] Configure Git appropriately for work
I use my work email when at work, so this configures that value.
Additionally, we use HTTPS instead of SSH to connect to Git repositories
because we require SAML tokens to access our private repositories. This
change overrides my usual preference of using SSH.

There's a slight wrinkle here because I'm now using my work tag on Spin
and it sets a macOS LaunchAgent file. I'm okay with this because Spin
environments are ephemeral anyway.