[emacs] Move Doom storage out of profile folder
This takes care of changing the location that Doom stores all of its
files. I need to set `DOOMLOCALDIR` to make the `doom` CLI programs take
my preference into account. It would be nice if `doom-local-dir` and
`doom-cache-dir` could be set in any Emacs file, but alas, that is not
the case.