[git] Restore autocompletion for git co
When I reworked my `git co` command to be agnostic to the name of the
main branch, I lost the ability to autocomplete branches. This change
relies on the `_git` completion built into `zsh-completion` (and usually
included by default in installations of `zsh`) to give the same
completion as `git checkout`.

I'm not wild about the reliance on something that isn't in my dotfiles,
but I also didn't want to just inline the implementation. Until I see
a problem, this is good enough for now.