[ruby] Adapt chruby to use .tool-versions as well
`.tool-versions` is the version file convention from [asdf][1] for
specifying versions of tools across multiple version managers. chruby
doesn't support this out-of-the-box, but modifying it to do so makes it
so I only ever have to use a single version manager file.

This isn't the cleanest, or probably fastest, way to do this, but I
think it's simple and fast enough that I'm okay with it. I can always
decide to optimize it later.

[1]: https://asdf-vm.com