[zsh] Use C-g to insert a Git branch name
I've lately been working on many different branches in a repository with
a large team. This causes me to lose the context for what my branch
names are. Due to this pain point, I now use fzf to find the Git branch
that I want with fuzzy searching.

My workflow is as follows:

    git co <C-g>

This checks out the branch after exiting the fuzzy finder.