[emacs] Replace isearch with Consult
Incremental search when using Evil is of limited utility because you end
up using `/` or `?` more often than `C-s` or `C-r`. However, having the
ability to see _all_ matches of a search in a file — and then act on
them — is useful.

This replaces `C-s` (`isearch-forward`) and `C-r` (`isearch-backward`)
with `consult-line` to make them more useful.