[ruby] Enable pry everywhere when using IRB
Pry is one of my favorite tools for Ruby. I use it on basically every
project I work on. Even when it's not explicitly configure, I like to
use it.

This change introduces an `IRBRC` file that attempts to load Pry and
start it when running IRB. If it's unable to load Pry, it continues with
the IRB session as normal.