[git] Add a helper for splitting commits by file
When working on chains of commits and rebasing them to tell a story,
it's often easiest to tell a story by isolating changes per file. This
change adds a commit that is able to split a commit into many commits by
file.

Note that if you're in a place where this is very useful, you have my
sympathies.