[mac] Add easy management for dark mode
In the summer, I'm finding myself getting headaches from the sun
streaming into the window behind my desk. After some experimentation,
I realized that this is due to using dark mode with the sunlight in the
background.

This change includes a new Hammerspoon spoon for managing dark mode in
macOS. The configuration also binds toggling dark mode to F18 on my
keyboard for easy access. That binding won't work when I'm laptop mode,
but I will cross that chasm when I get to it.

Additionally, this adds the configuration for Kitty that I have been
using for about a year. It's imperfect, but necessary for the use of the
dark mode configuration.