[emacs] Add a branch cleanup command to Magit
This change adds a `b #` command to my Magit dashboard that cleans up
merged local branches and deleted remote branches since it's likely that
I won't need either of those any more. This is another step to keep
myself in Emacs so I don't need to flip to the terminal as much.

This is a port of the logic for my `git cleanup` alias, but improved: it
doesn't show an error message when there aren't any deleted remote
branches.